Corrected final rule. Removing assembly/polishing directory failed when running the pipeline on multiple BACs

+++ Snakefile
@@ -127,8 +127,10 @@ rule CreateBlastDatabase:
     log:
         "log/{sample}_makeblastdb.log"
     threads: 1
-    conda: "envs/blast.yaml"
-    envmodules: "blast/2.9.0"
+    conda:
+        "envs/blast.yaml"
+#    envmodule:
+#        "blast/2.9.0"
     shell:
         """
         outbase={output}
@@ -152,7 +154,7 @@ rule AlignVector:
     log:
         "log/{sample}_blastVector.log"
     conda: "envs/blast.yaml"
-    envmodules: "blast/2.9.0"
+#    envmodules: "blast/2.9.0"
     threads: 1
     shell:
         """
@@ -179,7 +181,7 @@ rule AlignGeneA:
     log:
         "log/{sample}_blastGeneA.log"
     conda: "envs/blast.yaml"
-    envmodules: "blast/2.9.0"
+#    envmodules: "blast/2.9.0"
     threads: 1
     shell:
         """
@@ -206,7 +208,7 @@ rule AlignGeneB:
     log:
         "log/{sample}_blastGeneB.log"
     conda: "envs/blast.yaml"
-    envmodules: "blast/2.9.0"
+#    envmodules: "blast/2.9.0"
     threads: 1
     shell:
         """
@@ -233,7 +235,7 @@ rule Align2Host:
     log:
         "log/{sample}_hostalign.log"
     conda: "envs/minimap2.yaml"
-    envmodules: "minimap2/2.17"
+#    envmodules: "minimap2/2.17"
     threads: 4
     shell:
         """
@@ -255,7 +257,7 @@ rule Hostsam2paf:
         "align/minimap2/host/{sample}.paf"
     threads: 1
     conda: "envs/minimap2.yaml"
-    envmodules: "minimap2/2.17"
+#    envmodules: "minimap2/2.17"
     shell:
         """
         paftools.js sam2paf -p {input} > {output}
@@ -272,7 +274,7 @@ rule HostAlignStats:
         "align/minimap2/host/{sample}_alignment.stats"
     threads: 1
     conda: "envs/samtools.yaml"
-    envmodules: "samtools/1.9"
+#    envmodules: "samtools/1.9"
     shell:
         """
         TotalAlignments=$(samtools view -c {input})
@@ -308,7 +310,7 @@ rule Filter_HostAlignment_SAM2BAM:
         "align/minimap2/host/{sample}.bam"
     threads: 4
     conda: "envs/samtools.yaml"
-    envmodules: "samtools/1.9"
+#    envmodules: "samtools/1.9"
     shell:
         """
         if [ "{threads}" -ge 2 ];
@@ -345,7 +347,7 @@ rule Annotate_Reads:
         "tables/ReadClass/{sample}_ReadClass.rds"
     log:
         "log/{sample}_AnnotateReads.log"
-    envmodules: "r/3.6.0"
+#    envmodules: "r/3.6.0"
     singularity:
         "shub://pgpmartin/SingIMG:r36_nanobac"
     threads: 1
@@ -387,7 +389,7 @@ rule Select_VDVreads:
         outVDVnames = "SelectedReads/VDV/{sample}_Sel_VDV_ReadNames.tsv"
     log:
         "log/{sample}_Sel_VDV_Selection.log"
-    envmodules: "r/3.6.0"
+#    envmodules: "r/3.6.0"
     singularity:
         "shub://pgpmartin/SingIMG:r36_nanobac"
     threads: 1
@@ -420,7 +422,7 @@ rule selRead_Get_fastq:
     output:
         temp("SelectedReads/{readSelection}/{sample}_Sel_{readSelection}.fq")
     conda: "envs/seqtk.yaml"
-    envmodules: "seqtk/1.3"
+#    envmodules: "seqtk/1.3"
     threads: 1
     shell:
         """
@@ -470,7 +472,7 @@ rule Prepare_VDV_reads:
         VDVjunctions = "SelectedReads/VDVprepared/{sample}_VDVreads_InsertVectorJunctions.rds"
     log:
         "log/{sample}_VDVreadPreparation.log"
-    envmodules: "r/3.6.0"
+#    envmodules: "r/3.6.0"
     singularity:
         "shub://pgpmartin/SingIMG:r36_nanobac"
     threads: 4
@@ -504,7 +506,7 @@ rule RandomSampling_VDVreads:
     log:
         "log/{sample}_VDVreadsRandomSampling.log"
     conda: "envs/seqtk.yaml"
-    envmodules: "seqtk/1.3"
+#    envmodules: "seqtk/1.3"
     threads: 1
     shell:
         """
@@ -537,7 +539,7 @@ rule Kalign_RandomSetsOfVDVreads:
     log:
         "log/{sample}_kalign_RS{NUM}.log"
     conda: "envs/kalign2.yaml"
-    envmodules: "kalign/2.04"
+#    envmodules: "kalign/2.04"
     threads: 1
     shell:
         """
@@ -564,7 +566,7 @@ rule consensus_VDVrandomSets:
         consName = "cons_{sample}_RS{NUM}"
     log:
         "log/{sample}_ConsensusVDVRndSet_RS{NUM}.log"
-    envmodules: "r/3.6.0"
+#    envmodules: "r/3.6.0"
     singularity:
         "shub://pgpmartin/SingIMG:r36_nanobac"
     threads: 1
@@ -592,7 +594,7 @@ rule align_consensus_VDVrandomSets:
     log:
         "log/{sample}_consAlign.log"
     conda: "envs/kalign2.yaml"
-    envmodules: "kalign/2.04"
+#    envmodules: "kalign/2.04"
     threads: 1
     shell:
         """
@@ -619,7 +621,7 @@ rule consensus_fromVDVcons:
         consName = "cons_{sample}"
     log:
         "log/{sample}_ConsensusFromVDVcons.log"
-    envmodules: "r/3.6.0"
+#    envmodules: "r/3.6.0"
     singularity:
         "shub://pgpmartin/SingIMG:r36_nanobac"
     threads: 1
@@ -648,7 +650,7 @@ rule Select_longVDreads:
         outReadNames = "SelectedReads/longVD/{sample}_Sel_longVD_ReadNames.tsv"
     log:
         "log/{sample}_Sel_longVD_Selection.log"
-    envmodules: "r/3.6.0"
+#    envmodules: "r/3.6.0"
     singularity:
         "shub://pgpmartin/SingIMG:r36_nanobac"
     threads: 1
@@ -683,7 +685,7 @@ rule SelectSplit_DVDreads:
         outFasta = "SelectedReads/DVD/{sample}_SelSplit_DVD.fa"
     log:
         "log/{sample}_DVD_SelectAndSplit.log"
-    envmodules: "r/3.6.0"
+#    envmodules: "r/3.6.0"
     singularity:
         "shub://pgpmartin/SingIMG:r36_nanobac"
     threads: 1
@@ -736,7 +738,7 @@ rule align_VD_to_cons:
     log:
         "log/{sample}_polishing_{consensusType}.log"
     conda: "envs/minimap2.yaml"
-    envmodules: "minimap2/2.17"
+#    envmodules: "minimap2/2.17"
     threads: 4
     shell:
         """
@@ -758,7 +760,7 @@ rule racon_Polish:
     output:
         "assembly/polishing/{sample}_{consensusType}_polished.fa"
     conda: "envs/racon.yaml"
-    envmodules: "racon/1.4"
+#    envmodules: "racon/1.4"
     threads: 4
     shell:
         """
@@ -786,5 +788,10 @@ rule Polishing_step02:
     shell:
         """
         mv {input} {output}
-        rmdir assembly/polishing
+        if ls -1qA assembly/polishing/ | grep -q .
+        then
+          echo "assembly/polishing directory is not empty"
+        else
+          rmdir assembly/polishing
+        fi
         """
